# Conflicts:
#	lib/core/network/end_points.dart
#	lib/core/network/repository.dart
#	lib/core/util/cubit/cubit.dart
#	lib/core/util/cubit/state.dart
#	lib/features/account/pages/notificatio/notificatio.dart
#	lib/features/account/pages/notification/notification.dart
#	lib/features/saved/presentation/pages/saved.dart
#	lib/main.dart

+++ lib/features/category_details/presentation/widgets/cat_details_item.dart
@@ -97,7 +97,7 @@ class CatDetailsItem extends StatelessWidget {
                               }),
                           const Spacer(),
                           AvailableItem(
-                            amount: model.amount,
+                            amount: model.amount.toInt(),
                           ),
                         ],
                       ),
